requirements.txt
streamlit>=1.38
ultralytics==8.3.0
torch==2.4.1
torchvision==0.19.1
opencv-python-headless>=4.8
numpy
pillow
scikit-learn
matplotlib
git+https://github.com/jacobgil/pytorch-grad-cam.git
